src/speech_transcriber/requirements.txt
openai>=1.14
pydub>=0.25
python-dotenv
tqdm
reportlab
flask
